requirements.txt
langchain==0.2.14
langchain-community==0.2.12
openai>=1.54.0
httpx>=0.27.0
requests==2.32.3
python-dotenv==1.0.1
rank-bm25>=0.2.2
